### 2024-08-27 #7137 @AnnieKLamar
Hello there, thank you for the above information! Here are a few suggestions of reviewers that would be well suited to evaluate our submission:

- Corinna Coupette (Digital Humanities, Python, Networks)
- jkbren (Agent-Based Models, Network Science)
- Sam Hames (CS, DS, IS, Digital Humanities, Computational Humanities)
- Pietro Monitcone (CS, IS, network science)
- Moritz Makowski (Data pipelines, data generation, networks)
