### 2024-08-31 4869816
Update README.md
update to small-world control method call and epoch parameter name
