### 2024-09-10 c65a932
citation update
