+++ src/pythagora/experiments.py
@@ -330,17 +330,19 @@ def scale_free_simulation(number_of_buyers,
     for buyer in buyers:
         if buyer not in scale_free.nodes:
             remaining_buyers.append(buyer)
+    # PART 1: Add all buyers to market
     last = False
     while not last:
         set_to_last = False
         next_node_set = []
+        # this is the last set of buyers to add
         if len(remaining_buyers) < set_size or len(remaining_buyers) == set_size:
             for bu in remaining_buyers:
                 next_node_set.append(bu)
             set_to_last = True
-        else:
+        else:  # this is not the last set of buyers to add
             next_node_set = random.sample(remaining_buyers, set_size)
-        for b in next_node_set:
+        for b in next_node_set:  # remove added buyers from remaining buyers
             remaining_buyers.remove(b)
         scale_free = sf.build_scale_free_network(next_node_set, scale_free, community_bonus, log, epoch, buyers)
         scale_free = market.change_all_buyer_intentions(scale_free,
@@ -353,8 +355,14 @@ def scale_free_simulation(number_of_buyers,
             print("Epoch: " + str(epoch))
             print("Time elapsed: %s seconds " % timedelta(seconds=end_time - start_time))
         epoch += 1
+        if epoch >= number_of_epochs:
+            print("Target epochs reached, but buyers still remain outside market.")
+            print("Continuing experiment for minimum number of epochs required to add", number_of_buyers, "buyers to "
+                                                                                                          "market.")
         if set_to_last:
             last = True
+
+    # PART 2: Iterative buying until target epochs reached
     if epoch < number_of_epochs + 1:
         for e in range(epoch, number_of_epochs + 1):
             scale_free = market.change_all_buyer_intentions(scale_free,
