### 2024-09-10 f300a61
warning added to minimum epoch in scale free experiment; timestep reporting fixed for scale free experiment; testing complete
