### 2024-09-15 a5b792e
bug fix in logger.py
