### 2024-09-15 1d1f933
bug fix for null graphs created in time step 1 of SF control experiments
