### 2024-09-16 33cd580
bug fix for null graphs created in time step 1 of SF control experiments
