+++ src/pythagora/controls.py
@@ -48,7 +48,7 @@ def make_sf_control_graph_info_file(num_buyers,
     fields = ["experiment", "date", "time", "number of buyers", "initial_set_size",
               "set size", "min num communities", "min community fill",
               "assemblage", "total epochs", "upper threshold",
-              "lower threshold", "death threshold", "community bonus", "probability of link"]
+              "lower threshold", "death threshold", "community bonus"]
     writer.writerow(fields)
     info = ["scale free", str(date.today()), str((datetime.now()).strftime("%H:%M:%S")), str(num_buyers),
             str(initial_set_size), str(set_size), str(min_num_communities), str(min_community_fill),
@@ -75,9 +75,10 @@ def build_scale_free_control_network(set_of_nodes, G, community_bonus,
         current_node = set_of_nodes.pop()
         # add our node to the graph space
         t = len(G.nodes)
+        # move add_node line above t assignment to allow self loops
+        G.add_node(current_node)
         # assign probabilities to each node
         probs = {}  # (nodes, probabilities)
-
         if t != 0:
             for node in G.nodes():
                 if node.community == current_node.community:
@@ -90,12 +91,6 @@ def build_scale_free_control_network(set_of_nodes, G, community_bonus,
             for p in normalized_probs.keys():
                 w.append(normalized_probs[p])
             selected_node = random.choices(population, weights=w)
-
-            print(t)
-            print(current_node.name)
-            print(selected_node[0].name)
-            # move this line up to allow self loops
-            G.add_node(current_node)
             G.add_edge(current_node, selected_node[0])
     logger.log(G, epoch, 0, 0, all_buyers_in_market)
     return G
