### 2024-09-16 bb3e751
bug fix for edge additions in SF control
