### 2024-09-16 e621e4e
disallow self-loops from SF controls.py
