### 2024-10-03 #7137 @AnnieKLamar
@samhforbes Happy Thursday! Are you able to give us an estimated timeline for when this article will move to review or be assigned an editor?
