### 2024-11-08 #7137 @sbenthall
@jkbren @dataspider  Can you review this submission to JOSS?
