### 2024-11-09 #7137 @dataspider
@sbenthall Okay
@AnnieKLamar I skimmed the repo and the write-up, and it seems that you have a bunch of files in the repo that should have been gitignored, so please add a python .gitignore to your repo – GitHub has templates for that – and remove all the things that shouldn't have gotten into your repo in the first place. Also, you mention networkx in a specific version as a dependency, which probably should be `networkx >= 3.2` in your pyproject.toml (see [the pyproject.toml of this recent JOSS publication](https://github.com/mnets/pymnet/blob/master/pyproject.toml) as inspiration on how to specify dependencies in your pyproject.toml). It also seems that you don't have tests. In brief, I'd recommend you check the [JOSS Reviewer Guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html) and make sure that your submission conforms with the requirements listed therein, because reviewers will get this as a checklist and if some things are evidently missing, this will just prolong the review.
