### 2024-11-11 a882cff
updated and moved .gitignore file; added networkx to .toml requirements
