### 2024-11-11 #7137 @AnnieKLamar
@dataspider Hi there, thanks for quick response!
- I have added all the contents of the .idea folder to the .gitignore file.
- I have added NetworkX 3.2 to the .toml file requirements.
- We included "manual steps described so that the functionality of the software can be verified" in the QuickStart instructions--are the automated tests required as well?
