### 2024-11-12 #7137 @sbenthall
Thanks @dataspider for agreeing to review, and for your comments here!

JOSS has a somewhat more formal review process, which involves a checklist. That process will begin once we find another reviewer.

I'll add you as a first reviewer -- you might want to review the guidelines:
https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Please let me know if you have any questions or concerns about this, especially if this comes to you as a surprise.
