### 2024-11-12 #7137 @sbenthall
@alanlujan91 Would you be interested in and able to review this submission to JOSS?
