### 2024-11-22 #7137 @sbenthall
@pitmonticone would you be willing to review this submission to JOSS?
