### 2024-11-25 #7137 @dataspider
@sbenthall Could you share your thoughts on [this issue](https://github.com/openjournals/joss/issues/1395#issuecomment-2484114962)? I am aware that this submission was previously on scope review (just as the other submission I reference there, which I am now handling separately with the editor and AEiC of that submission), and the concerns I detail in the linked issue are even more pertinent to this present submission.
