### 2024-11-25 #7137 @sbenthall
Thanks for your question @dataspider

I believe @samhforbes removed the 'query-scope' label, indicating that he believed this submission is in scope as is.

But I'm happy to discuss the scope issue before beginning formal review if you have concerns.

Which criterion do you think this submission fails? Is it this:

"- Be feature-complete (no half-baked solutions) and _be designed for maintainable extension (not one-off modifications)_."

?

What indicators in the code are the basis of this evaluation?
