### 2024-12-02 #7549 @editorialbot
[REVIEW]: Pythagora: A Python Package for Modeling the Impact of Social Networks on Market Outcomes
**Submitting author:** @AnnieKLamar (<a href="http://orcid.org/0000-0002-0562-9444">Annie K. Lamar</a>)
**Repository:** https://github.com/stwilker/pythagora
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.0
