### 2024-12-02 #7549 @sbenthall
@AnnieKLamar One issue that has been flagged by the editorial bot is that the license of your submission (Attribution-NonCommercial-ShareAlike 4.0 International) is not OSI (Open Source Initiative) approved.

I wonder if you would consider an alternative license for your submission. I can also ask the other editors if there are any possible exemptions to the OSI rule.
