### 2024-12-02 #7137 @dataspider
@sbenthall My understanding is that once I get the checklist, there is no way to reject the submission as out of scope, which is precisely why I opened the issue that was now closed. Imho the submission is problematic regarding both of the criteria I quoted, and I don't want to have to sign off on something that doesn't meet these criteria. I understand that JOSS wants authors to work with reviewers to meet the checklist criteria, but when meeting the scope criteria would essentially require a complete revamp...? In any event, if you want to proceed with the review as in-scope, I would kindly ask you to assign this to someone else, and I'll make sure to do my own internal scope review before accepting any review requests in the future.
