### 2024-12-02 #7137 @sbenthall
@dataspider Sorry for the confusion. I didn't get a response from you earlier, so I thought perhaps you had withdrawn your objection.

It sounds like you are asking the editors to do a scope review.

@samhforbes I see that you earlier assigned this submission for editorial review, but then later removed that label before you assigned me as editor.

Did an editorial review of this submission happen? Why did you remove that label?

I thought you had reviewed the scope and found it to be within scope. Am I mistaken?

---------------

My own opinion: I don't see at the moment how this submission is obviously out of scope. I'm curious what @dataspider would want to see to see the package more convincingly display substantial scholarly effort.

One thing I see is a lack of automated tests, which is on the reviewer checklist.

Another thing is that the paper has only 4 references, which perhaps does not place it well within the broader range of literature and tools.  I.e, the  'State of the field: Do the authors describe how this software compares to other commonly-used packages?' requirement, which is on the reviewers checklist.
