### 2024-12-02 #7137 @sbenthall
Thanks @samhforbes for that clarification.

It looks like the editors have determined this submission to be in scope!

Does that change your views on reviewing this submission @dataspider ?

As you point out, the purpose of JOSS reviews is to improve the submission, and that's even more true when there's constructive feedback to be given. Since scope was determined by the editors to not be the issue here, I wonder what the other underlying issues are. Is it possible to articulate those? You would be welcome to raise those in your review, even if they are not on the checklist.
