### 2024-12-02 #7137 @dataspider
Let me start by articulating that "substantial scholarly effort" is a terrible name for a checklist criterion (who am I to judge what may have been a substantial effort for others).
I guess my underlying intuition is that this feels more like a move to get a quick additional publication with associated citations than like a software contribution. While this is subjective and might be misguided (let's assume everyone has heard best intentions), there are more objective problems that would make it highly problematic for this software to be used by others for doing simulations. For example, unless I missed something, none of the randomness is seeded, which makes the results computationally unreproducible. From the code it's pretty apparent that no attempt was made to follow elementary software-engineering principles. This is widespread (although still problematic) for the throwaway code people write for their one-off publications, but unacceptable for software "stamped" to be reused by others, especially in fields with low computational literacy. It simply creates huge research-validity problems. Again, I am not in principle opposed to helping others improve their software, but this just seems more like a start-from-scratch than an evolutionary improvement to what's there. Does that help?
