### 2024-12-03 #7137 @sbenthall
@dataspider Thanks for clarifying your position.

As a purely procedural matter, it seems to me that:
 - the scope check was asked for
 - the editors determined the submission was in scope

There are two ways to proceed:
 - (A) we continue on to the formal review, which is typically when reviewers assess the submission and recommend, or not, acceptance into the journal. It appears you are well on your way to doing that already.
 - (B) you withdraw as a reviewer (because you disagree with the editors about scope), and we look for a replacement reviewer.

It's up to you. Would you prefer (A) or (B)?
