### 2024-12-03 #7137 @dataspider
@sbenthall Let's do (B) then. If I simulate (A) (assuming away the scope issue), then my conclusion is "major revision", but I don't want to hold up the acceptance process.
