### 2024-12-03 #7549 @sbenthall
@AnnieKLamar I expect you have been following the conversation on #7137

@dataspider has requested to be removed as a reviewer. I will soon begin a search for another one.

I want to apologize for the procedural challenges here. This has been an irregular process.

I wonder if you have any response to @dataspider's recommendations before I try to restart the review process.

In particular, it has been proposed that you might resubmit the package, but with 'major revisions'. Would you be open to that? Or would you prefer to continue on with the current submission for formal review?
