### 2024-12-04 #7549 @AnnieKLamar
Hi @sbenthall -- thank you for managing this process.

Given that this package is the result of several years of work and -- to my understanding -- meets the checklist requirements, we do not think that our work requires major revisions. We will continue with the current submission.

We do have a couple questions:

 1. We would like to make small changes -- would you let us know if it is okay to make these changes within the context of the current submission, rather than re-submitting?

- Change the license type to one that is OSI approved. We apologize for the oversight.
- Add more detail to the previous scholarship / references section, which we intentionally left sparse for the sake of brevity and the context of JOSS, but we are happy to expand to clarify the positionality of our work.

2. The former reviewer seemed especially concerned with the "Automated Tests" checklist item ("Are there automated tests or manual steps described so that the functionality of the software can be verified?"). We included manual steps and extensive documentation, which we believed met this requirement (based on our reading of several previously published JOSS papers). Do you have a suggestion or clarification in regards to this item?

Thank you again for your time managing the review process--we appreciate your feedback.
