### 2024-12-19 #7549 @sbenthall
@dostuffthatmatters We we looking for a second reviewer for this JOSS submission, to fill in for one that has dropped off. You were mentioned by the authors as a potential reviewer for this piece. Would you be willing to take it on?
