### 2024-12-19 #7549 @AnnieKLamar
@sbenthall

It's clear we need to significantly re-think the framing of our paper and augment our code. We would like to withdraw this submission and re-submit after these corrections are made. Sincere thanks for your time and the advice of the reviewers in this thread.
