### 2024-12-19 #7549 @sbenthall
@samhforbes Please see this thread. The authors have withdrawn the submission. From the docs, it looks like only the EIC can issue the `withdraw` command to the editorialbot. Thanks.
